+++ app/controllers/principal_reviewer/data_requests_controller.rb
@@ -13,6 +13,7 @@ class PrincipalReviewer::DataRequestsController < ApplicationController
   def review
     original_status = @data_request.status
     original_dataset_ids = @data_request.dataset_ids.sort
+
     if AgreementTransaction.save_agreement!(@data_request, current_user, request.remote_ip, "agreement_update", data_request_params: data_request_params)
       if original_status != "approved" && @data_request.status == "approved"
         @data_request.save_signature!(:reviewer_signature_file, params[:data_uri]) if params[:data_uri].present?
@@ -21,8 +22,6 @@ class PrincipalReviewer::DataRequestsController < ApplicationController
         @data_request.sent_back_for_resubmission_email(current_user)
       elsif original_status != "closed" && @data_request.status == "closed"
         @data_request.close_daua!(current_user)
-      elsif original_status != "expired" && @data_request.status == "expired"
-        @data_request.expire_daua!(current_user)
       end
       @data_request.compute_datasets_added_removed!(original_dataset_ids, current_user)
       respond_to do |format|